Try to implement gkrm on x86_64. Currently implementing gkmm and gkrm using Goto algorithm is problematic due to the temporary rank-kc update in type TV.
Given that the output c may have different types than the rank-kc update.
The worst case is to allocate an m-by-n buffer. Several things are pretty
sure:

1. allocating this buffer may break user's program due to the size.
2. allocating V may result in significant overhead.

GPU gkmx does not have this problem because the temporary rank-kc update is
stored in the registers. However, gkrm still need to allocate an m-by-n/4
buffer to perform global reduction.